Handle undefined `block_number` (#1288)
This PR adds additional logic to handle rare cases of undefined block_number.

There are certain rare cases, where response from mirror-node does not have transaction_index or block_number properties. That can happen for example when contracts/results api is called before all the data is saved in the DB. A single re-fetch should fix this. But in case this is not enough, we can now check if block_number is undefined and return null. This is okay for all ethereum tools, as in ethereum block_number == undefined means that transaction is pending. So even if we return null, we are not braking any usecases.

---------

Signed-off-by: georgi-l95 <glazarov95@gmail.com>