Improve `eth_sendRawTransaction` latency by simplifying precheck (#1446)
This PR reduces eth_sendRawTransaction latency by reducing the number of requests and simplifying precheck.
Now we make 1 less request to the mirror node and 1 less request to consensus node. This reduces both hbar cost and latency. Performance test was done and metrics we collected by prometheus, by average transactions now take 10% less.

Also remove the FileAppendTransaction getReceipt, as transaction in hedera are executed in the order they are received, so we know that the next one will be executed after this one, no need to wait.

Also remove the getReceipt for file

Signed-off-by: georgi-l95 <glazarov95@gmail.com>