Fix `eth_estimateGas` transaction values formatting (#1615)
This PR fixes estimateGas, because in some cases it throws Unable to parse JSON. Ethereum tools are passing value and gasPrice fields as hex, but mirror-node is expecting int. To fix this we need to format properly this fields. Additional checks and more careful formatting is needed for the value to avoid losing accuracy.

---------

Signed-off-by: georgi-l95 <glazarov95@gmail.com>